fix float32 print on some mcu's for now

+++ src/mcu_utils/basictypes.nim
@@ -98,20 +98,22 @@ proc repr*(ts: UBits64): string =
 
 basicMathBorrows(Volts)
 fdivMathBorrows(Volts)
+import std/strformat
+
 proc repr*(ts: Volts): string =
-  return $(ts.float32) & "'V "
+  return $(ts.int32) & "'V "
 
 basicMathBorrows(Amps)
 fdivMathBorrows(Amps)
 proc repr*(ts: Amps): string =
-  return $(ts.float32) & "'A "
+  return $(ts.int32) & "'A "
 
 basicMathBorrows(Volts64)
 fdivMathBorrows(Volts64)
 proc repr*(ts: Volts64): string =
-  return $(ts.float32) & "'V "
+  return $(ts.int32) & "'V "
 
 basicMathBorrows(Amps64)
 fdivMathBorrows(Amps64)
 proc repr*(ts: Amps64): string =
-  return $(ts.float32) & "'A "
+  return $(ts.int32) & "'A "
